Move editor components to separate folder/module
