+++ src/components/projecteditor/editors/editor-contract.js
@@ -18,7 +18,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import style from './style-editor-contract.less';
-import { IconTrash, IconAdd } from '../icons';
+import { IconTrash, IconAdd } from '../../icons';
 
 class ConstructorArgument extends Component {
