+++ src/components/projecteditor/editors/index.js
@@ -1,0 +1,4 @@
+export { default as AccountEditor } from  './editor-account';
+export { default as AppEditor } from  './editor-app';
+export { default as ContractEditor } from  './editor-contract';
+export { default as Editor } from  './editor';
